+++ src/C++/Message.cpp
@@ -247,8 +247,8 @@ std::string& Message::toString( std::string& str,
                                 int bodyLengthField, 
                                 int checkSumField ) const
 {
-  int length = bodyLength( beginStringField, bodyLengthField, checkSumField );
-  m_header.setField( IntField(bodyLengthField, length) );
+  size_t length = bodyLength( beginStringField, bodyLengthField, checkSumField );
+  m_header.setField( IntField(bodyLengthField, (int)length) );
   m_trailer.setField( CheckSumField(checkSumField, checkSum(checkSumField)) );
 
 #if defined(_MSC_VER) && _MSC_VER < 1300
@@ -601,7 +601,7 @@ void Message::validate() const
     const BodyLength& aBodyLength = FIELD_GET_REF( m_header, BodyLength );
 
     const int expectedLength = (int)aBodyLength;
-    const int actualLength = bodyLength();
+    const int actualLength = (int)bodyLength();
 
     if ( expectedLength != actualLength )
     {
